+++ frontend/frontend-api/src/main/java/org/bloomreach/forge/settings/management/config/LoadableDetachableConfigModel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2019 BloomReach Inc. (http://www.bloomreach.com)
+ * Copyright 2013-2020 Bloomreach Inc. (http://www.bloomreach.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,11 +26,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * @author Jeroen Reijn
+ * LoadableDetachableConfigModel
  */
 public abstract class LoadableDetachableConfigModel<T> extends LoadableDetachableModel<T> {
 
-    private static final long serialVersionUID = 1L;
     private final Logger log = LoggerFactory.getLogger(this.getClass());
 
     public CMSFeatureConfig getConfig() {