+++ frontend/frontend-core/src/main/java/org/bloomreach/forge/settings/management/config/eventlog/EventLogConfigPanel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2019 BloomReach Inc. (http://www.bloomreach.com)
+ * Copyright 2013-2020 Bloomreach Inc. (http://www.bloomreach.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,7 +28,6 @@ import org.bloomreach.forge.settings.management.config.CMSFeatureConfig;
 
 
 public class EventLogConfigPanel extends FeatureConfigPanel {
-    private static final long serialVersionUID = 1L;
     private EventLogConfigModel detachableEventLogConfig = new EventLogConfigModel();
 
     public EventLogConfigPanel(IPluginContext context, IPluginConfig config) {