+++ frontend/frontend-core/src/main/java/org/bloomreach/forge/settings/management/config/passwordpolicies/ChangePasswordConfigModel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2019 BloomReach Inc. (http://www.bloomreach.com)
+ * Copyright 2013-2020 Bloomreach Inc. (http://www.bloomreach.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,11 +19,10 @@ package org.bloomreach.forge.settings.management.config.passwordpolicies;
 import org.bloomreach.forge.settings.management.config.LoadableDetachableConfigModel;
 
 /**
- * @author Jeroen Reijn
+ * ChangePasswordConfigModel
  */
 public class ChangePasswordConfigModel extends LoadableDetachableConfigModel<ChangePasswordConfig> {
 
-    private static final long serialVersionUID = 1L;
     public static final String CHANGE_PASSWORD_CONFIG_PATH = "/hippo:configuration/hippo:frontend/cms/cms-dashshortcuts/changePasswordShortcut";
 
     @Override