+++ frontend/frontend-core/src/main/java/org/bloomreach/forge/settings/management/config/upload/AssetUploadSettingConfigModel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2019 BloomReach Inc. (http://www.bloomreach.com)
+ * Copyright 2013-2020 Bloomreach Inc. (http://www.bloomreach.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,10 +22,8 @@ import org.bloomreach.forge.settings.management.config.LoadableDetachableConfigM
  * Loadable detachable model of the image upload config.
  * Configuration can be found in the repository at path:
  * /hippo:configuration/hippo:workflows/gallery/asset-gallery/frontend:renderer
- * @author Jeroen Reijn
  */
 public class AssetUploadSettingConfigModel extends LoadableDetachableConfigModel<UploadSettingsConfig> {
-    private static final long serialVersionUID = 1L;
     public static final String ASSET_UPLOAD_CONFIG_PATH = "/hippo:configuration/hippo:workflows/gallery/asset-gallery/frontend:renderer";
 
     @Override