+++ frontend/frontend-core/src/main/java/org/bloomreach/forge/settings/management/config/upload/AssetValidationServiceConfigModel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2019 BloomReach Inc. (http://www.bloomreach.com)
+ * Copyright 2013-2020 Bloomreach Inc. (http://www.bloomreach.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,11 +22,9 @@ import org.bloomreach.forge.settings.management.config.LoadableDetachableConfigM
  * Loadable detachable model of the asset upload config.
  * Configuration can be found in the repository at path:
  * hippo:configuration/hippo:frontend/cms/cms-services/assetValidationService
- * @author Jeroen Reijn
  */
 public class AssetValidationServiceConfigModel extends LoadableDetachableConfigModel<AssetValidationServiceConfig> {
 
-    private static final long serialVersionUID = 1L;
     public static final String ASSET_VALIDATION_CONFIG_PATH = "/hippo:configuration/hippo:frontend/cms/cms-services/assetValidationService";