+++ frontend/frontend-core/src/main/java/org/bloomreach/forge/settings/management/config/upload/ImageValidationServiceConfigModel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2019 BloomReach Inc. (http://www.bloomreach.com)
+ * Copyright 2013-2020 Bloomreach Inc. (http://www.bloomreach.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,10 +22,8 @@ import org.bloomreach.forge.settings.management.config.LoadableDetachableConfigM
  * Loadable detachable model of the image upload config.
  * Configuration can be found in the repository at path:
  * hippo:configuration/hippo:frontend/cms/cms-services/imageValidationService
- * @author Jeroen Reijn
  */
 public class ImageValidationServiceConfigModel extends LoadableDetachableConfigModel<ImageValidationServiceConfig> {
-    private static final long serialVersionUID = 1L;
     public static final String IMAGE_VALIDATION_CONFIG_PATH = "/hippo:configuration/hippo:frontend/cms/cms-services/imageValidationService";
 
     @Override