+++ frontend/frontend-core/src/main/java/org/bloomreach/forge/settings/management/config/upload/ImageValidationServiceConfigPanel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2019 BloomReach Inc. (http://www.bloomreach.com)
+ * Copyright 2013-2020 Bloomreach Inc. (http://www.bloomreach.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -37,7 +37,6 @@ import org.bloomreach.forge.settings.management.config.CMSFeatureConfig;
  * Panel for configuring the image validation service.
  */
 public class ImageValidationServiceConfigPanel extends FeatureConfigPanel {
-    private static final long serialVersionUID = 1L;
     private ImageValidationServiceConfigModel imageValidationServiceConfigModel = new ImageValidationServiceConfigModel();
 
     public ImageValidationServiceConfigPanel(IPluginContext context, IPluginConfig config) {
@@ -73,7 +72,6 @@ public class ImageValidationServiceConfigPanel extends FeatureConfigPanel {
         // Don't do this in a listview
         final ListView<String> allowedExtensionsView;
         allowedExtensionsView = new ListView<String>("imagevalidation-allowedExtensions", new PropertyModel(imageValidationServiceConfig, "allowedExtensions")) {
-            private static final long serialVersionUID = 1L;
             @Override
             protected void populateItem(final ListItem<String> item) {
                 RequiredTextField extensionField = new RequiredTextField("image-extension", item.getModel());
@@ -81,8 +79,6 @@ public class ImageValidationServiceConfigPanel extends FeatureConfigPanel {
                 item.add(extensionField);
 
                 AjaxSubmitLink remove = new AjaxSubmitLink("remove") {
-                    private static final long serialVersionUID = 1L;
-
                     @Override
                     protected void onSubmit(AjaxRequestTarget target, Form form) {
                         getModelObject().remove(item.getModelObject());
@@ -95,7 +91,6 @@ public class ImageValidationServiceConfigPanel extends FeatureConfigPanel {
         };
 
         AjaxLink addExtension = new AjaxLink("add-extension") {
-            private static final long serialVersionUID = 1L;
 
             @Override
             public void onClick(AjaxRequestTarget target) {