+++ frontend/frontend-core/src/main/java/org/bloomreach/forge/settings/management/config/usermanagement/UserManagementConfigModel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2019 BloomReach Inc. (http://www.bloomreach.com)
+ * Copyright 2013-2020 Bloomreach Inc. (http://www.bloomreach.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,11 +19,10 @@ package org.bloomreach.forge.settings.management.config.usermanagement;
 import org.bloomreach.forge.settings.management.config.LoadableDetachableConfigModel;
 
 /**
- * @author Jeroen Reijn
+ * Model for UserManagementConfig
  */
 public class UserManagementConfigModel extends LoadableDetachableConfigModel<UserManagementConfig> {
 
-    private static final long serialVersionUID = 1L;
     public static final String USER_MANAGEMENT_CONFIG_MODEL_PATH = "/hippo:configuration/hippo:frontend/cms/cms-admin/users";
 
     @Override