+++ src/site/markdown/release-notes.md
@@ -1,5 +1,5 @@
 <!--
-  Copyright 2013-2019 BloomReach Inc. (https://www.bloomreach.com)
+  Copyright 2013-2020 Bloomreach Inc. (https://www.bloomreach.com)
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -16,11 +16,21 @@
 
 ## Release notes
 
+### Version 3.0.0 
+Release date: 29 January 2020
+
++ Upgrade to Bloomreach Experience Manager 14, including the demo project.<br/>
+  The "Other" section (tab) was removed since it contained only settings for the Social Sharing plugin, which was
+  removed from version 14.   
++ The translations of configured tabs moved from hippo:translation subnodes below<br/>
+ ```/hippo:configuration/hippo:frontend/cms/cms-admin/settings/tabconfig``` to <br/>
+ ```/hippo:configuration/hippo:translations/hippo:cms/settingsmanagement```.
+
 ### Version 2.0.0 
 Release date: 2 March 2019
 
 + [HIPFORGE-278](https://issues.onehippo.com/browse/HIPFORGE-278)<br/> 
-  Upgrade to BloomReach CMS 13, including demo project. 
+  Upgrade to Bloomreach Experience Manager 13, including the demo project. 
   This also entails renaming of Maven coordinates and Java packaging from 'onehippo' to 'bloomreach'. 
   When upgrading, please revisit the [installation page](install.html).
 + [HIPFORGE-271](https://issues.onehippo.com/browse/HIPFORGE-271)<br/>
@@ -30,11 +40,10 @@ Release date: 2 March 2019
 + [HIPFORGE-280](https://issues.onehippo.com/browse/HIPFORGE-280)<br/>
   Make add/remove form data exclude paths work.
 
-
 ### Version 1.0.0 
 Release date: 26 Feb 2018
 
-  Upgrade to BloomReach CMS 12, including demo project. 
+  Upgrade to Bloomreach CMS 12, including demo project. 
 + Textual improvements, in Dutch mainly.
 
 ### Version 0.6.0